pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "cae-cli"
version = "1.0.0"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
description = "SolidWorks CAE集成助手 - 专业的CAE命令行工具"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Manufacturing",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: CAD",
    "Topic :: Scientific/Engineering :: Mechanical Engineering",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Environment :: Console",
]
keywords = ["cae", "cad", "solidworks", "freecad", "fea", "engineering", "cli"]

dependencies = [
    "click>=8.0.0",
    "rich>=13.0.0",
    "numpy>=1.21.0",
    "pyyaml>=6.0",
    "jinja2>=3.0.0",
    "pint>=0.22",
    "requests>=2.25.0",
    "markdown>=3.0.0",
]

[project.optional-dependencies]
# AI功能（本地向量库+模型）
ai = [
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
]
# 完整功能（包含几何处理）
full = [
    "pythonocc-core>=7.7.0",
    "meshio>=5.0.0",
    "vtk>=9.0.0",
    "matplotlib>=3.5.0",
    "pandas>=1.3.0",
]
# 仅网格分析
tools = [
    "meshio>=5.0.0",
    "matplotlib>=3.5.0",
]
# 优化功能（包含绘图和数据可视化）
optimize = [
    "matplotlib>=3.5.0",
    "numpy>=1.21.0",
]
# SSH增强功能
ssh = [
    "paramiko>=3.0.0",
]

# GUI依赖
gui = [
    "PySide6>=6.5.0",
    "PySide6-Addons>=6.5.0",
    "PySide6-WebEngine>=6.5.0",
]

# 开发依赖
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "pyinstaller>=6.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/cae-cli"
Documentation = "https://cae-cli.readthedocs.io"
Repository = "https://github.com/yourusername/cae-cli.git"
Issues = "https://github.com/yourusername/cae-cli/issues"

[project.scripts]
# 主命令
cae-cli = "sw_helper.cli:cli"
# 别名
sw-helper = "sw_helper.cli:cli"
# GUI 命令
cae-gui = "main_gui:main"

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["sw_helper*"]
exclude = ["tests*"]

[tool.setuptools.package-data]
sw_helper = ["data/*", "templates/*"]

# 开发工具配置
[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --capture=no"

[tool.coverage.run]
source = ["src/sw_helper"]
omit = ["*/tests/*", "*/test_*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
